+++ src/main/resources/changelog.md
@@ -4,6 +4,7 @@
 - Chiseled Sandstone Slab & Chiseled Sandstone Stairs
 - missing recipes (stonecutting & smelting)
 - blasting recipes
+- missing advancements to unlock recipes
 
 #### Change :
 - h